Extract out the JSON tests to share between MySQL/Postgres
